Mem_System1/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
pymysql==1.1.0
cryptography==41.0.7
neo4j==5.16.0
qdrant-client==1.7.0

# Vector & Embeddings
openai==1.10.0
numpy==1.26.3
sentence-transformers==2.3.1

# NLP & Text Processing
nltk==3.8.1
tiktoken==0.5.2
rank-bm25==0.2.2
jieba==0.42.1

# Graph Processing
networkx==3.2.1
python-louvain==0.16

# HTTP Client
requests==2.31.0
aiohttp==3.9.1
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
PyYAML==6.0.1
colorama==0.4.6

# Data Processing
pandas==2.1.4
scipy==1.11.4

# Optional: LightRAG (if using lightrag engine)
# lightrag @ git+https://github.com/HKUDS/LightRAG.git

# Optional: Development
black==23.12.1
pytest==7.4.3
pytest-asyncio==0.23.3

# Optional: Monitoring
prometheus-client==0.19.0
